element modal: merge edit and show modals
