+++ app/modules/listings/components/listing_element_actions.svelte
@@ -2,33 +2,13 @@
   import { clone } from 'underscore'
   import { icon } from '#app/lib/icons'
   import { moveArrayElement } from '#app/lib/utils'
-  import Modal from '#components/modal.svelte'
   import Spinner from '#general/components/spinner.svelte'
-  import { removeElement, updateElement, askUserConfirmationAndRemove } from '#listings/lib/listings'
-  import ElementEditor from '#modules/listings/components/element_editor.svelte'
+  import { updateElement } from '#listings/lib/listings'
   import { i18n } from '#user/lib/i18n'
 
-  export let isReordering, element, elements, listingId, flash
+  export let isReordering, element, elements, flash, index
 
-  let isLoading, isEditMode
-  const { _id, uri } = element
-  $: hasSeveralElements = elements.length > 1
-  $: index = elements.findIndex(obj => obj.uri === uri)
-
-  async function onRemoveElement (e) {
-    const element = e.detail
-    await askUserConfirmationAndRemove(_removeElement, element?.comment)
-      .catch(err => flash = err)
-  }
-
-  async function _removeElement () {
-    return removeElement(listingId, uri)
-      .then(() => {
-        // Enhancement: after remove, have an "undo" button
-        elements.splice(index, 1)
-        elements = elements
-      })
-  }
+  let isLoading
 
   async function reorder (newIndex) {
     isReordering = isLoading = true
@@ -38,7 +18,7 @@
     const exclusiveOrdinal = newIndex + 1
 
     await updateElement({
-      id: _id,
+      id: element._id,
       ordinal: exclusiveOrdinal,
     })
       .then(() => {
@@ -53,59 +33,37 @@
       })
   }
 
-  function toggleEditMode () {
-    isEditMode = !isEditMode
-  }
+  $: hasSeveralElements = elements.length > 1
 </script>
-{#if isEditMode}
-  <Modal on:closeModal={() => isEditMode = false}
-  >
-    <ElementEditor
-      bind:element
-      on:editorDone={toggleEditMode}
-      on:removeElement={onRemoveElement}
-    />
-  </Modal>
-{/if}
-
-<div class="element-actions">
-  <button
-    class="edit-button tiny-button"
-    on:click={toggleEditMode}
-  >
-    {@html icon('pencil')}
-  </button>
-  {#if hasSeveralElements}
-    <div class="reorder-section">
-      {#if index !== 0}
-        <button
-          disabled={isReordering}
-          on:click={() => reorder(index - 1)}
-          title={i18n('Move up')}
-        >
-          {@html icon('chevron-up')}
-        </button>
-      {/if}
-      <div class="list-index">
-        {#if isLoading}
-          <Spinner />
-        {:else}
-          {index + 1}
-        {/if}
-      </div>
-      {#if index !== elements.length - 1}
-        <button
-          disabled={isReordering}
-          on:click={() => reorder(index + 1)}
-          title={i18n('Move down')}
-        >
-          {@html icon('chevron-down')}
-        </button>
+{#if hasSeveralElements}
+  <div class="reorder-section">
+    {#if index !== 0}
+      <button
+        disabled={isReordering}
+        on:click={() => reorder(index - 1)}
+        title={i18n('Move up')}
+      >
+        {@html icon('chevron-up')}
+      </button>
+    {/if}
+    <div class="list-index">
+      {#if isLoading}
+        <Spinner />
+      {:else}
+        {index + 1}
       {/if}
     </div>
-
-  {/if}
-</div>
+    {#if index !== elements.length - 1}
+      <button
+        disabled={isReordering}
+        on:click={() => reorder(index + 1)}
+        title={i18n('Move down')}
+      >
+        {@html icon('chevron-down')}
+      </button>
+    {/if}
+  </div>
+{/if}
 <style lang="scss">
   @import "#general/scss/utils";
   .reorder-section{
@@ -113,16 +71,6 @@
     button{
       color: $grey;
     }
-  }
-  .element-actions{
-    @include display-flex(row, center);
-    margin: 0.3em;
-  }
-  .edit-button{
-    margin: 1em;
-    padding: 0.3em;
-    color: $grey;
-    background-color: $off-white;
   }
   .list-index{
     font-size: 1.3em;
